requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.3
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2022.1
requests==2.28.1
semantic-version==2.10.0
setuptools-rust==1.4.1
sqlparse==0.4.2
twilio==7.12.0
typing-extensions==4.3.0
urllib3==1.26.11
